Work on threading and porting audio layer to use pulseaudio.
This is work on the threading model for the program.  It was
originally desinged to run on a small microporocessor where the
program would haved exclusive use of system resources.  This
isn't true of the UDRC/Raspberry Pi.  We created two extra threads
to handle the load.  One to handle Audio reading/writing, one to
handle the receive processing and outgoing serial data, and one to
handle transmit processing and incoming serial data.

Also ported the audio layer to use portaudio instead.  This will
allow for easier split-channel usage on the UDRC which will allow
for uses such as a dual-stack repeater, or a repeater and an APRS
digi run off of one computing device.  This should also save some
processing in mmdvm-udrc because it shouldn't have to do resampling
or conversion to float from int.

Work is ongoing, this commit probably doesn't make anything work.